Add files via upload
 a cocktail drink search engine

create an input search that adds an event listener to a click while passing in a function

declare constant, link its value to that of a url's json.

create a fetch for url, using .then to parse response as JSON, then assign selectors with appropriate innerText/src data, with a .catch to prevent errors.

enjoy? enjoy.