Add files via upload
Using a fetch to grab src data from a card deck website, juxtaposing 2 declared random card values, essentially recreating the game War! 

Using an if loop to print the winner (or a tie) to the console based on values

Interesting issue found. Each local session is a full deck. One can only play as many times as there are cards, so 26 times. Beyond that, there is a 'no cards left' returned on the console. Will think on this.